+++ packages/ispect/lib/src/features/log_viewer/presentation/screens/logs_screen.dart
@@ -62,6 +62,7 @@ class _LogsScreenState extends State<LogsScreen> {
     _logsViewController = widget.controller ??
         ISpectViewController(
           onShare: widget.options.onShare,
+          metadataProvider: widget.options.metadataProvider,
           initialSettings: widget.options.initialSettings,
           onSettingsChanged: _handleSettingsChanged,
         );